Load recipes after the item repository has finished importing
